Submodule update

+++ src/chain.cpp
@@ -12,7 +12,7 @@
 
 namespace SPOA {
 
-std::unique_ptr<Chain> createChain(uint32_t id, const char* name, uint32_t name_length,
+std::unique_ptr<Chain> createChain(uint64_t id, const char* name, uint32_t name_length,
     const char* data, uint32_t data_length) {
 
     assert(name_length);
@@ -21,7 +21,7 @@ std::unique_ptr<Chain> createChain(uint32_t id, const char* name, uint32_t name_
     return std::unique_ptr<Chain>(new Chain(id, name, name_length, data, data_length));
 }
 
-std::unique_ptr<Chain> createChain(uint32_t id, const char* name, uint32_t name_length,
+std::unique_ptr<Chain> createChain(uint64_t id, const char* name, uint32_t name_length,
     const char* data, uint32_t data_length, const char* quality, uint32_t quality_length) {
 
     assert(name_length);
@@ -32,12 +32,12 @@ std::unique_ptr<Chain> createChain(uint32_t id, const char* name, uint32_t name_
         quality, quality_length));
 }
 
-Chain::Chain(uint32_t id, const char* name, uint32_t name_length, const char* data,
+Chain::Chain(uint64_t id, const char* name, uint32_t name_length, const char* data,
     uint32_t data_length)
         : id_(id), name_(name, name_length), data_(data, data_length), quality_() {
 }
 
-Chain::Chain(uint32_t id, const char* name, uint32_t name_length, const char* data,
+Chain::Chain(uint64_t id, const char* name, uint32_t name_length, const char* data,
     uint32_t data_length, const char* quality, uint32_t quality_length)
         : id_(id), name_(name, name_length), data_(data, data_length), quality_(quality, quality_length) {
 }
